added download button for spiderman image
note: had to do a work around to be able to download images. HTML blocks access to downloading files saved on a local machine, so the image 'crossOrigin' must be set to 'anonymous' and the image src must be an http link to work around this. seems safer this way too